Added new tests for the App, Modal and horizontalList components
Using Jest created two new tests to check if the Modal and the App component render correctly.
Also added a test to check if the horizontalList component renders correctly, and if the data is passed correctly to the component.
For App and Modal only used the test to render the component and check if it does correctly, for the horizontalList we need to give the component data to render, and see if it renders correctly.

Co-authored-by: Antonio3veces3 <aantonioramirez33@gmail.com>
Co-authored-by: J3fte <santillanabdiel0@gmail.com>
Co-authored-by: Axw3ll <axelomar.a.v@gmail.com>